requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
pymongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.2
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
urllib3==2.5.0
